Move command handler into bot folder
